게시물 수정, 삭제 권한 체크

+++ src/main/java/com/koreaIT/java/am/ArticleDeleteServlet.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
+import java.util.Map;
 
 import com.koreaIT.java.am.config.Config;
 import com.koreaIT.java.am.util.DBUtil;
@@ -14,13 +15,15 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 @WebServlet("/article/doDelete")
 public class ArticleDeleteServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		
+	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
+
 		response.setContentType("text/html; charset=utf-8");
 		Connection conn = null;
 
@@ -30,15 +33,51 @@ public class ArticleDeleteServlet extends HttpServlet {
 			Class.forName(Config.getDBDriverName());
 			String url = Config.getDBUrl();
 			conn = DriverManager.getConnection(url, Config.getDBUser(), Config.getDBPwd());
-			
-			int id = Integer.parseInt(request.getParameter("id"));//넘겨 받은 id 형변환.
-			
+
+			int id = Integer.parseInt(request.getParameter("id"));// 넘겨 받은 id 형변환.
+
 			SecSql sql = new SecSql();
-			sql.append("DELETE FROM article");
+			sql.append("SELECT *");
+			sql.append("FROM article");
 			sql.append("WHERE id =?", id);
+
+			Map<String, Object> article = DBUtil.selectRow(conn, sql);
+
+			if (article.isEmpty()) {
+				response.setContentType("text/html; charset=utf-8");
+				response.getWriter().append(
+						String.format("<script>alert('%d번 게시물은 존재하지 않습니다.'); location.replace('list');</script>", id));
+				return;
+			}
+
+			HttpSession session = request.getSession();
+			int loginedMemberId = -1;
+
+			if (session.getAttribute("loginedMemberId") != null) {
+				loginedMemberId = (int) session.getAttribute("loginedMemberId");
+			}
+
+			if (loginedMemberId == -1) {
+				response.setContentType("text/html; charset=UTF-8");
+				response.getWriter().append(
+						String.format("<script>alert('로그인 후 이용해주세요'); location.replace('../member/login');</script>"));
+				return;
+			}
+
+			if ((int) article.get("memberId") != loginedMemberId) {
+				response.setContentType("text/html; charset=UTF-8");
+				response.getWriter().append(String.format(
+						"<script>alert('해당 게시물에 대한 권한이 없습니다'); location.replace('detail?id=%d');</script>", id));
+				return;
+			}
+
+			sql = new SecSql();
+			sql.append("DELETE FROM article");
+			sql.append("WHERE id = ?", id);
 			DBUtil.delete(conn, sql);
-			response.getWriter().append(String.format("<script>alert('%d번 게시물 삭제 성공'); location.replace('list');</script>", id));
-		
+			response.getWriter()
+					.append(String.format("<script>alert('%d번 게시물 삭제 성공'); location.replace('list');</script>", id));
+
 		} catch (ClassNotFoundException e) {
 			System.out.println("드라이버 로딩 실패");
 		} catch (SQLException e) {
